% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{dist_bar_chart}
\alias{dist_bar_chart}
\title{Functions to plot bar chart}
\usage{
dist_bar_chart(data, index_var_names, index_var_levels, title)
}
\arguments{
\item{data}{input data generated from \code{dist_bar_chart_dataPrep} function}

\item{index_var_names}{Index variable}

\item{index_var_levels}{Contents of each level in \code{index_var}}

\item{title}{Stacked bar chart title}
}
\value{
Returns stacked bar chart to demonstrate the distribution of \code{index_var} at each level of \code{grouping_var} (indicated in \code{dist_bar_chart_dataPrep} function)
}
\description{
\code{dist_bar_chart} return stacked bar chart
}
\details{
This function allows users to generate stacked bar chart visualization
}
\keyword{visualization}
