% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{dist_bar_chart_dataPrep}
\alias{dist_bar_chart_dataPrep}
\title{Functions to prepare data for visualization}
\usage{
dist_bar_chart_dataPrep(data, grouping_var, index_var)
}
\arguments{
\item{data}{input data containing both \code{grouping_var} and \code{index_var}}

\item{grouping_var}{Grouping variable}

\item{index_var}{Index variable}
}
\value{
Returns the value of \code{output}
}
\description{
\code{dist_bar_chart_dataPrep} return data to be used in generate visualization (using \code{dist_bar_chart}).
}
\details{
This function allows users to prepare data for stacked bar chart visualization
}
\keyword{visualization}
