% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientMetricHW-package.R
\docType{package}
\name{GradientMetrics-package}
\alias{GradientMetrics}
\alias{GradientMetrics-package}
\title{GradientMetrics: My take on the assignment}
\description{
This package analyses the expermient and survey data.
}
\author{
\strong{Maintainer}: Lucas Enrich \email{lucas.a.enrich@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
