% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_func.R
\name{model_performance_metrics}
\alias{model_performance_metrics}
\title{index_func}
\usage{
model_performance_metrics(predictions, targets)
}
\arguments{
\item{predictions}{predicted values of the response variable}

\item{targets}{actual values of the response variable}
}
\value{
Returns the value of \code{output}
}
\description{
\code{model_performance_metrics} return the accuracy, precision, recall and F1 score based on model predictions and target values
}
\details{
A Model Performance Evaluation Function

This function allows users to calculate accuracy, precision, recall and F1 score based on model predictions and target values
}
