% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{philosophy_m1_bar_chart}
\alias{philosophy_m1_bar_chart}
\title{Functions to plot stacked bar chart for Philosophy Module 1 questions}
\usage{
philosophy_m1_bar_chart(data, qs_levels)
}
\arguments{
\item{data}{input data}

\item{qs_levels}{content of question}
}
\value{
Returns stacked bar chart to demonstrate the distribution of answers in Philosophy Module 1 questions
}
\description{
\code{philosophy_m1_bar_chart} return stacked bar chart of answer
}
\details{
This function allows users to generate stacked bar chart to illustrate the distribution of answers in Philosophy Module 1 questions
}
\keyword{visualization}
