% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{pie_chart}
\alias{pie_chart}
\title{Functions to plot pie chart}
\usage{
pie_chart(data, var, var_levels, var_level_names)
}
\arguments{
\item{data}{input data}

\item{var}{Variable to examine its distribution}

\item{var_levels}{Levels in \code{var}}

\item{var_level_names}{Level names for \code{var}}
}
\value{
Returns pie chart to demonstrate the distribution of \code{var} in survey data
}
\description{
\code{pie_chart} return pie chart of \code{var}
}
\details{
This function allows users to generate pie chart to illustrate the distribution of \code{var} in survey data
}
\keyword{visualization}
