% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratified_experimient}
\alias{stratified_experimient}
\title{Stratified sample of experimient}
\format{
A data frame with 6000 rows and 9 variables:
\describe{
\item{response_id}{unique id of respondanse}
\item{task}{unknown}
\item{duration}{duration of promotion. Value of experimient}
\item{offer}{which was the offer. Value of experimient}
\item{outcome}{promised outcome. Value of experimient}
\item{price}{price offered. Value of experimient}
\item{rtb}{how the offer was phrased. Value of experimient}
\item{social_proof}{What legitimizes our product. Value of experimient}
\item{answer}{price offered. Value of experimient}
...
}

A data frame with 6000 rows and 9 variables:
\describe{

...
}
}
\source{
\url{https://github.com/lucariel/GradientMetrics/data}

\url{https://github.com/lucariel/GradientMetrics/data}
}
\usage{
stratified_experimient

stratified_experimient
}
\description{
A dataset containing 1500 samples of each answer (1,2,3,4).
}
\keyword{datasets}
